+++ apps/cold-graphql/src/backend/entities/hooks/product-tag-assignment.hooks.ts
@@ -1,0 +1,28 @@
+
+// ProductTagAssignment Hooks
+import { CreateOrUpdateHookParams, ReadHookParams, DeleteHookParams } from '@exogee/graphweaver';
+import { BaseSidecar } from '../base.sidecar';
+import { OrgContext } from '../../libs/acls/acl_policies';
+import { ProductTagAssignment } from '../postgresql';
+
+export class ProductTagAssignmentHooks extends BaseSidecar {
+	constructor() {
+		super(ProductTagAssignment, 'product_tag_assignments');
+	}
+	// Overrride BeforeReadHook here:
+
+	// Overrride AfterReadHook here:
+
+	// Overrride BeforeCreateHook here:
+
+	// Overrride AfterCreateHook here:
+
+	// Overrride BeforeUpdateHook here:
+
+	// Overrride AfterUpdateHook here:
+
+	// Overrride BeforeDeleteHook here:
+
+	// Overrride AfterDeleteHook here:
+
+}